requirements.txt
tensorflow
matplotlib
numpy
pandas
Pillow
scikit-learn
seaborn
scipy
tqdm
torch
torchvision
timm
grad-cam